Fix tool timer overflow on Windows, bundle JRE for Freerouting
- agent_frame.cpp: use %lld/long long for tool elapsed timer timestamp
  (long is 32-bit on MSVC, truncating the 64-bit ms timestamp)
- kicad_cli_util.h: check for bundled JRE at bin/freerouting/jre/bin/java.exe
  before falling back to system PATH on Windows

+++ agent/agent_frame.cpp
@@ -196,7 +196,7 @@ static wxString BuildRunningToolHtml( int aIndex, const wxString& aDesc )
         "class=\"tool-result-header py-2.5 px-3 flex items-center gap-2\">"
         "<span class=\"text-text-secondary text-[12px]\">%s</span>"
         "<span class=\"tool-status text-text-muted text-[12px] ml-auto flex items-center gap-2\">"
-        "<span class=\"tool-elapsed\" data-start=\"%ld\"></span>"
+        "<span class=\"tool-elapsed\" data-start=\"%lld\"></span>"
         "<a href=\"agent:cancel_tool\" style=\"color:var(--text-muted); font-size:11px; "
         "text-decoration:none; opacity:0.7; cursor:pointer;\" "
         "onmouseover=\"this.style.opacity='1';this.style.color='var(--accent-red)'\" "
@@ -208,7 +208,7 @@ static wxString BuildRunningToolHtml( int aIndex, const wxString& aDesc )
         "data-toggle-type=\"toolresult\" data-toggle-index=\"%d\" style=\"display:none;\">"
         "</div>"
         "</div>",
-        aIndex, aDesc, (long) wxGetUTCTimeMillis().GetValue(), aIndex );
+        aIndex, aDesc, (long long) wxGetUTCTimeMillis().GetValue(), aIndex );
 }
